order-delivery-time-handler/requirements.txt
kafka-python
black
httpx
mpu
uszipcode
pytest
pytest-cov
tox
requests_mock
python-dotenv
python-Levenshtein
psycopg2-binary
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-exporter-jaeger